+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- [Added missing attributes of transaction types](https://github.com/multiversx/mx-sdk-dapp/pull/613)
 
 ## [[v2.4.0](https://github.com/multiversx/mx-sdk-dapp/pull/604)] - 2023-02-03