+++ CHANGELOG.md
@@ -7,8 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 - [Added missing attributes of transaction types](https://github.com/multiversx/mx-sdk-dapp/pull/613)
-
 - [Add Opera Crypto wallet provider - Beta](https://github.com/multiversx/mx-sdk-dapp/pull/600)
+- [Add default SignTransactionModal for extra provider](https://github.com/multiversx/mx-sdk-dapp/pull/612)
+
 ## [[v2.4.0](https://github.com/multiversx/mx-sdk-dapp/pull/604)] - 2023-02-03
 
 - [Add support for webview provider automatic login](https://github.com/multiversx/mx-sdk-dapp/pull/609)